Port the code from #29228
